refactor: Phase 4 — extract ModelQueryBuilder into its own module
- Move ModelQueryBuilder class (~590 lines) from Ad4mModel.ts to ModelQueryBuilder.ts
- Uses import type { Ad4mModel } to avoid circular runtime dependency
- Ad4mModel.ts query() method still returns ModelQueryBuilder via runtime import
- Re-exports from Ad4mModel.ts and barrel index.ts preserve public API
- Ad4mModel.ts reduced from ~2860 to ~2260 lines
- All 119 tests pass (115 passed, 4 skipped)
